Add files via upload
As can be seen from the pictures, firstly, you have to choose the location of each city by clicking your mouse. After that, the GA measures the optimal destination between cities, and the salesperson will find the shortest possible route to visit each city exactly once and returns to the origin city. Crossover and Mutation operations are done based on the location of the cities and the most valuable routes will be identified.